+++ tests/testthat/test-ggsave.R
@@ -56,6 +56,21 @@ test_that("ggsave can handle blank background", {
   expect_true(grepl("fill: none", bg))
 })
 
+test_that("ggsave warns about empty or multiple filenames", {
+  filenames <- c("plot1.png", "plot2.png")
+  plot <- ggplot(mtcars, aes(disp, mpg)) + geom_point()
+
+  withr::with_file(filenames, {
+    expect_warning(
+      suppressMessages(ggsave(filenames, plot)),
+      "`filename` must have length 1"
+    )
+    expect_error(
+      ggsave(character(), plot),
+      "`filename` cannot be empty."
+    )
+  })
+})
 
 # plot_dim ---------------------------------------------------------------